AI-Agent-Client/requirements.txt
requests==2.31.0\
rich==13.7.1\
langchain==0.1.16\
python-dotenv==1.0.1}